Primeira tentativa de trazer o menu para a tela
