+++ src/pages/Hall/index.jsx
@@ -1,16 +1,11 @@
 import React from 'react'
-import { Link, useHistory } from 'react-router-dom';
+import { useHistory } from 'react-router-dom';
 import Button from '@material-ui/core/Button';
-import Typography from '@material-ui/core/Typography';
 import Container from '@material-ui/core/Container';
-import {useStyles, Logo, SingIn, Copyright, NavBar} from '../../components.js';
-import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos';
-
-
+import {useStyles, NavBar} from '../../components.js';
 import '../../style.css';
 
 function Home(){
-  
 
   const classes = useStyles();
   const history = useHistory();
@@ -23,22 +18,6 @@ function Home(){
     history.push('/Menu')
   }
 
-
-    
-  function getLocalStorage(){
-      const tokenLocal = localStorage.getItem('token')
-      console.log(tokenLocal)
-
-      fetch('https://lab-api-bq.herokuapp.com/products', {
-        method: 'GET',
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/x-www-form-urlencoded',
-          'Authorization': tokenLocal ,
-        },
-      }).then((response) => console.log(response.json()))
-  }
-
     return (
       <React.Fragment >
         <NavBar/>
@@ -49,8 +28,7 @@ function Home(){
               </Button>
               <Button type="submit" fullWidth variant="contained" className={classes.submitHall} onClick={routerPendentes} >
                 Pendentes
-              </Button>
-              <Button onClick={(event) => { event.preventDefault(); getLocalStorage();}}>Produtos</Button>
+              </Button>           
             </Container>   
         </React.Fragment>
